Move from double to single precision to save memory

+++ src/fft.c
@@ -26,15 +26,15 @@
  * returns:
  *   an array of size `length` transformed from `input` OR NULL if OOM
  */
-double complex * calc_fft(double complex input[], int length) {
-    double complex *twiddles = (double complex *)Heap_Calloc(length / 2 * sizeof(double complex));
+float complex * calc_fft(float complex input[], int length) {
+    float complex *twiddles = (float complex *)Heap_Calloc(length / 2 * sizeof(float complex));
     if(twiddles==NULL){
         return NULL;
     }
     for(int i=0;i<length/2;i++){
-        double complex J = 2.0 * I;
+        float complex J = 2.0 * I;
         *(twiddles + i) =
-            cexp(((double complex)M_PI * (double complex)i * J) / (double complex)length);
+            cexp(((float complex)M_PI * (float complex)i * J) / (float complex)length);
     }
     return _fft(input, twiddles, length, length);
     Heap_Free(twiddles);
@@ -47,8 +47,8 @@ double complex * calc_fft(double complex input[], int length) {
  *   length, length of the original array
  *   origlength, length used to comptue the twiddle factors
  */
-double complex* _fft(double complex input[], double complex twiddles[], int length, int origlength) {
-    double complex *result = (double complex*)Heap_Calloc(length * sizeof(double complex));
+float complex* _fft(float complex input[], float complex twiddles[], int length, int origlength) {
+    float complex *result = (float complex*)Heap_Calloc(length * sizeof(float complex));
 
     if(length % 2){
         return dft(input, length);
@@ -56,8 +56,8 @@ double complex* _fft(double complex input[], double complex twiddles[], int leng
 
     int halflength = floor(length / 2);
 
-    double complex *input_even = (double complex*)Heap_Calloc(halflength * sizeof(double complex));
-    double complex *input_odd = (double complex*)Heap_Calloc(halflength * sizeof(double complex));
+    float complex *input_even = (float complex*)Heap_Calloc(halflength * sizeof(float complex));
+    float complex *input_odd = (float complex*)Heap_Calloc(halflength * sizeof(float complex));
 
     for(int i=0;i<length;i++) {
         if(i % 2){
@@ -67,8 +67,8 @@ double complex* _fft(double complex input[], double complex twiddles[], int leng
         }
     }
 
-    double complex *output_even = _fft(input_even, twiddles, halflength, origlength);
-    double complex *output_odd = _fft(input_odd, twiddles, halflength, origlength);
+    float complex *output_even = _fft(input_even, twiddles, halflength, origlength);
+    float complex *output_odd = _fft(input_odd, twiddles, halflength, origlength);
 
     for(int k=0;k<halflength;k++){
         *(result+k) = *(result+k) +
@@ -97,8 +97,8 @@ double complex* _fft(double complex input[], double complex twiddles[], int leng
  * returns:
  *   an array of size `length` transformed from `input` 
  */
-double complex* dft(double complex input[], int length) {
-    double complex *result = (double complex *)Heap_Calloc(length * sizeof(double complex));
+float complex* dft(float complex input[], int length) {
+    float complex *result = (float complex *)Heap_Calloc(length * sizeof(float complex));
 
     if(length == 1){
         *(result) = input[0];
@@ -106,23 +106,23 @@ double complex* dft(double complex input[], int length) {
         *(result) = (input[0] + input[1]);
         *(result+1) = (input[0] - input[1]);
     }else if(length == 4){
-        double complex a = (input[0] + input[2]);
-        double complex b = (input[0] - input[2]);
-        double complex c = (input[1] + input[3]);
-        double complex d = (input[1] - input[3]);
+        float complex a = (input[0] + input[2]);
+        float complex b = (input[0] - input[2]);
+        float complex c = (input[1] + input[3]);
+        float complex d = (input[1] - input[3]);
 
         *(result) = a + c;
         *(result+1) = b + (1.0 * I * d);
         *(result+2) = a - c;
         *(result+3) = b - (1.0 * I * d);
     }else{
-        double complex *twiddles = (double complex *)Heap_Calloc(length * sizeof(double complex));
+        float complex *twiddles = (float complex *)Heap_Calloc(length * sizeof(float complex));
         // Calculate the first half of the twiddle factors
         for(int k=0;k<(floor(length/2)+1);k++){
         //     twiddles = [math.e**(inv*2j*math.pi*k/N) for k in xrange(M)]+[N]
-            double complex J = 2.0 * I;
+            float complex J = 2.0 * I;
             *(twiddles + k) =
-                cexp(((double complex)M_PI * (double complex)k * J) / (double complex)length);
+                cexp(((float complex)M_PI * (float complex)k * J) / (float complex)length);
         }
         // Second hald of the twiddle factors are the conjugates of the first half
         for(int k=(floor(length/2)+1);k<length;k++){
